requirements.txt
requests
pandas
scikit-learn
apache-airflow
dvc
dvc-gdrive